requirements.txt
Django==5.0
djangorestframework==3.14
djangorestframework-simplejwt==5.2.0
python-decouple==6.1
